+++ Simulation_engine/exceptions.py
